Fix another bunch of errors

+++ modules/core/perf/src/perf.cpp
@@ -34,7 +34,7 @@ void ppc::core::Perf::task_run(const std::shared_ptr<PerfAttr>& perfAttr,
 void ppc::core::Perf::common_run(const std::shared_ptr<PerfAttr>& perfAttr, const std::function<void()>& pipeline,
                                  const std::shared_ptr<ppc::core::PerfResults>& perfResults) {
   auto begin = std::chrono::high_resolution_clock::now();
-  for (int i = 0; i < perfAttr->num_running; i++) {
+  for (uint64_t i = 0; i < perfAttr->num_running; i++) {
     pipeline();
   }
   auto end = std::chrono::high_resolution_clock::now();
